// kinetic-web/src/shared/api/types.ts
export interface components {
    schemas: {
// ...
        MeDto: {
            /** Format: uuid */
            id: string;
            email: string;
            firstName: string;
            lastName: null | string;
            avatarUrl: null | string;
            currentWorkspace: null | components["schemas"]["WorkspaceDto"];
        };
// ...
        Response: {
            message: string;
        };
        ResponseOfListOfProjectDto: {
            data?: null | components["schemas"]["ProjectDto"][];
            message?: string;
        };
        ResponseOfListOfProjectMemberDto: {
            data?: null | components["schemas"]["ProjectMemberDto"][];
            message?: string;
        };
// ...
        ResponseOfProjectDto: {
            data?: components["schemas"]["ProjectDto"];
            message?: string;
        };
// ...
    };
    responses: never;
    parameters: never;
    requestBodies: never;
    headers: never;
    pathItems: never;
}